Fixed error zip in PlayerInfoActivity (onErrorResumeNext( )) Add LeakCanary in App Added check Internet Connection to TabSearch and PlayerInfoActivity(including fragments) BottomNavigationBehavior (remove show/hide animate, changed display rules SnackBar )

+++ app/src/main/java/com/example/albert/dotasearch/repository/SearchPlayersRepository.java
@@ -7,8 +7,12 @@ import android.util.Log;
 import com.example.albert.dotasearch.App;
 import com.example.albert.dotasearch.database.AppDatabase;
 import com.example.albert.dotasearch.model.FavoritePlayer;
+import com.example.albert.dotasearch.model.FoundPlayer;
 import com.example.albert.dotasearch.util.UtilDota;
 
+import java.io.IOException;
+import java.net.InetSocketAddress;
+import java.net.Socket;
 import java.util.List;
 
 import io.reactivex.Single;
@@ -19,13 +23,15 @@ import io.reactivex.schedulers.Schedulers;
 public class SearchPlayersRepository {
     private LiveData<List<FavoritePlayer>> allFavoritePlayer;
     private AppDatabase db;
-    public static final String TAG = "FavoriteRepository";
-    private MutableLiveData<Boolean> isRequestSearchSuccessful;
+    public static final String TAG = "SearchPlayersRepository";
+    private MutableLiveData<Integer> responseStatusCode;
+    private static final int NO_INTERNET_CODE = -100;
+    private static final int ERROR_RX = -200;
 
     public SearchPlayersRepository() {
         db = App.get().getDB();
         allFavoritePlayer = db.favoritePlayerDao().getAllRx();
-        isRequestSearchSuccessful = new MutableLiveData<>();
+        responseStatusCode = new MutableLiveData<>();
     }
 
     public LiveData<List<FavoritePlayer>> getAllFavoritePlayer() {
@@ -45,20 +51,55 @@ public class SearchPlayersRepository {
     }
 
     public void searchResult(final String query) {
-        Log.d(TAG, "NETWORK");
-        Disposable dis = UtilDota.initRetrofitRx().getFoundPlayersRx(query)
-                .subscribeOn(Schedulers.io())
+        Disposable disposable = hasInternetConnection().flatMap(internet -> {
+            if (internet) {
+
+                Log.d(TAG, "NETWORK");
+                Disposable dis = UtilDota.initRetrofitRx().getFoundPlayersRxResponse(query)
+                        .subscribeOn(Schedulers.io())
+                        .subscribe(
+                                response -> {
+                                    List<FoundPlayer> foundPlayers = response.body();
+                                    db.foundPlayerDao().deleteAllFoundPlayer();
+                                    db.foundPlayerDao().insertAll(foundPlayers);
+                                    responseStatusCode.postValue(response.code());
+                                },
+                                err -> {
+                                    err.printStackTrace();
+                                    responseStatusCode.postValue(ERROR_RX);
+                                });
+            } else {
+                responseStatusCode.postValue(NO_INTERNET_CODE);
+            }
+
+            return Single.just(internet);
+        }).subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(
-                        response -> {
-                            db.foundPlayerDao().deleteAllFoundPlayer();
-                            db.foundPlayerDao().insertAll(response);
-                            isRequestSearchSuccessful.postValue(true);
-                        },
-                        err -> Log.e(TAG, err.getLocalizedMessage())
+                        internet -> Log.d(TAG, "internet is " + internet),
+                        Throwable::printStackTrace
                 );
     }
 
-    public MutableLiveData<Boolean> getIsRequestSearchSuccessful() {
-        return isRequestSearchSuccessful;
+    public MutableLiveData<Integer> getResponseStatusCode() {
+        return responseStatusCode;
+    }
+
+    public Single<Boolean> hasInternetConnection() {
+        return Single.fromCallable(() -> {
+            try {
+                // Connect to Google DNS to checkValidateTeamsData for connection
+                int timeoutMs = 1500;
+                Socket socket = new Socket();
+                InetSocketAddress socketAddress = new InetSocketAddress("8.8.8.8", 53);
+
+                socket.connect(socketAddress, timeoutMs);
+                socket.close();
+
+                return true;
+            } catch (IOException io) {
+                return false;
+            }
+        });
     }
 }
